backend/requirements.txt
fastapi
uvicorn
pydantic
beautifulsoup4

openai
python-dotenv
httpx
